requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.3
mercadopago==2.3.0
packaging==26.0
psutil==7.2.2
psycopg2-binary==2.9.11
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
python-dotenv==1.2.2
requests==2.33.1
setuptools==82.0.1
urllib3==2.6.3
Werkzeug==3.1.8
gunicorn==23.0.0
